isolate scatteringMaterial.c source code
