+++ pytissueoptics/rayscattering/opencl/src/propagation.c
@@ -1,5 +1,6 @@
 #include "random.c"
 #include "vectorOperators.c"
+#include "scatteringMaterial.c"
 
 
 void decreaseWeightBy(__global photonStruct *photons, float delta_weight, uint gid){
@@ -19,25 +20,6 @@ void moveBy(__global photonStruct *photons, float distance, uint gid){
     photons[gid].position += (distance * photons[gid].direction);
 }
 
-float getScatteringDistance(__global float * randomNums, float mu_t, uint gid){
-    return -log(randomNums[gid]) / mu_t;
-}
-
-float getScatteringAnglePhi(__global float * randomNums, uint gid){
-    float phi = 2.0f * M_PI * randomNums[gid];
-    return phi;
-}
-
-float getScatteringAngleTheta(__global float * randomNums, float g, uint gid){
-    if (g == 0){
-        return acos(2.0f * randomNums[gid] - 1.0f);
-    }
-    else{
-        float temp = (1.0f - g * g) / (1 - g + 2 * g * randomNums[gid]);
-        return acos((1.0f + g * g - temp * temp) / (2 * g));
-    }
-}
-
 void scatterBy(__global photonStruct *photons, float phi, float theta, uint gid){
     rotateAroundAxisGlobal(&photons[gid].er, &photons[gid].direction, phi);
     rotateAroundAxisGlobal(&photons[gid].direction, &photons[gid].er, theta);
@@ -53,32 +35,9 @@ void roulette(__global photonStruct *photons, __global uint * randomSeedBuffer, 
     }
 }
 
-
 bool getIntersection(float distance) {
     return false;
 }
-
-// move to scatteringMaterial.c
-struct ScatteringAngles {
-    float phi, theta;
-};
-
-typedef struct ScatteringAngles ScatteringAngles;
-
-ScatteringAngles getScatteringAngles(uint gid,
-           __global photonStruct *photons, __constant materialStruct *materials,
-           __global float *randomNums, __global uint *seedBuffer)
-{
-    ScatteringAngles angles;
-    randomNums[gid] = getRandomFloatValue(seedBuffer, gid);
-    angles.phi = getScatteringAnglePhi(randomNums, gid);
-    randomNums[gid] = getRandomFloatValue(seedBuffer, gid);
-    float g = materials[photons[gid].material_id].g;
-    angles.theta = getScatteringAngleTheta(randomNums, g, gid);
-    return angles;
-}
-
-
 
 void scatter(uint gid, uint logIndex,
            __global photonStruct *photons, __constant materialStruct *materials, __global loggerStruct *logger,
@@ -130,19 +89,4 @@ __kernel void propagate(uint dataSize, float weightThreshold,
                                 photons, materials, logger, randomNums, seedBuffer);
         stepIndex++;
     }
-}
-
-__kernel void getScatteringDistanceKernel(__global float * distanceBuffer, __global float * randomNums, float mu_t){
-    uint gid = get_global_id(0);
-    distanceBuffer[gid] = getScatteringDistance(randomNums, mu_t, gid);
-}
-
-__kernel void getScatteringAnglePhiKernel(__global float * angleBuffer,  __global float * randomNums){
-    uint gid = get_global_id(0);
-    angleBuffer[gid] = getScatteringAnglePhi(randomNums, gid);
-}
-
-__kernel void getScatteringAngleThetaKernel(__global float * angleBuffer,  __global float * randomNums, float g){
-    uint gid = get_global_id(0);
-    angleBuffer[gid] = getScatteringAngleTheta(randomNums, g, gid);
 }
